blast away deprecated pipe-shell
