+++ particles/PipeApps/ArtistAutofill.arcs
@@ -1,6 +1,13 @@
+<<<<<<< HEAD:particles/PipeApps/ArtistAutofill.arcs
 import './schemas/Json.arcs'
 import '../Services/schemas/RandomData.arcs'
 import '../Services/particles/Random.arcs'
+=======
+import './Json.schema'
+import './PipeEntity.schema'
+import '../Services/schemas/RandomData.schema'
+import '../Services/particles/Random.particle'
+>>>>>>> blast away deprecated pipe-shell:particles/PipeApps/ArtistAutofill.recipes
 
 schema Artist
   Text type
@@ -14,11 +21,17 @@ particle RandomArtist in './source/RandomArtist.js'
 particle SuggestArtist in './source/SuggestArtist.js'
   in Artist artist
   out Json suggestion
-  consume app
+  consume content
 
-recipe Com_spotify_music &autofill
+particle RequireQuery in './source/Noop.js'
+  in PipeEntity query
+
+recipe AutofillSpotifyMusic &artist_autofill
   create as artist
   create as suggestion
+  use #artist_autofill as query
+  RequireQuery
+    query = query
   RandomParticle
     randomData -> randomArtist
   RandomArtist
@@ -27,3 +40,4 @@ recipe Com_spotify_music &autofill
   SuggestArtist
     artist = artist
     suggestion = suggestion
+  description `suggest artist`
