+++ particles/PipeApps/PipeApps.recipes
@@ -1,0 +1,6 @@
+import 'MapsAutofill.recipes'
+import 'PeopleAutofill.recipes'
+import 'ArtistAutofill.recipes'
+import 'Caption.recipes'
+import 'UpdateMe.recipes'
+import 'ArtistPipe.recipes'
