+++ particles/PipeApps/Trigger.arcs
@@ -1,6 +1,6 @@
 import './schemas/PipeEntity.arcs'
 
-particle Trigger in './source/Noop.js'
+particle Trigger in './source/Trigger.js'
   in PipeEntity pipe
   consume root
-    provide app
+    provide content
