+++ particles/PipeApps/UpdateMe.recipes
@@ -1,4 +1,4 @@
-import './schemas/Json.arcs'
+import './Json.schema'
 
 particle UpdateMe in './source/UpdateMe.js'
   out Json output
