+++ particles/PipeApps/source/SuggestPerson.js
@@ -13,7 +13,7 @@
 defineParticle(({DomParticle, html, log}) => {
   return class extends DomParticle {
     get template() {
-      return html`<div slotid="assistance"></div>`;
+      return html`<span></span>`;
     }
     update({recentEntities}, state) {
       if (recentEntities) {
