--- particles/PipeApps2/BackgroundPipes.arcs
@@ -1,9 +1,0 @@
-import 'schemas/PipeEntity.arcs'
-
-particle BackgroundPipes in './source/Noop.js'
-  in [PipeEntity] entities
-
-recipe BackgroundPipes
-  create #pipeEntities as entities
-  BackgroundPipes
-    entities = entities
