--- particles/PipeApps2/MapsAutofill.arcs
@@ -1,21 +1,0 @@
-import './schemas/PipeEntity.arcs'
-import './schemas/Json.arcs'
-
-particle SuggestAddress in './source/SuggestAddress.js'
-  in [PipeEntity] recentEntities
-  out Json suggestion
-  consume content
-
-particle RequireQuery in './source/Noop.js'
-  in PipeEntity query
-
-recipe AutofillMaps &address_autofill
-  use #address_autofill as query
-  map 'pipe-entities' as recentEntities
-  create as suggestion
-  RequireQuery
-    query = query
-  SuggestAddress
-    recentEntities = recentEntities
-    suggestion = suggestion
-  description `suggest address`
