--- particles/PipeApps2/Trigger.arcs
@@ -1,6 +1,0 @@
-import './schemas/PipeEntity.arcs'
-
-particle Trigger in './source/Trigger.js'
-  in PipeEntity pipe
-  consume root
-    provide content
