--- particles/PipeApps2/schemas/Json.arcs
@@ -1,2 +1,0 @@
-schema Json
-  Text json
