--- particles/PipeApps2/schemas/PipeEntity.arcs
@@ -1,7 +1,0 @@
-schema PipeEntity
-  Text id
-  Text type
-  Text name
-  Number timestamp
-  Number count
-  Text source
