--- particles/PipeApps2/source/Trigger.js
@@ -1,19 +1,0 @@
-/**
- * @license
- * Copyright 2019 Google LLC.
- * This code may only be used under the BSD style license found at
- * http://polymer.github.io/LICENSE.txt
- * Code distributed by Google as part of this project is also
- * subject to an additional IP rights grant found at
- * http://polymer.github.io/PATENTS.txt
- */
-
-'use strict';
-
-defineParticle(({DomParticle, html}) => {
-  return class extends DomParticle {
-    get template() {
-      return html`<div slotid="content"></div>`;
-    }
-  };
-});
