--- shells/pipes-shell/node/serve.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-# node --no-warnings --experimental-modules --loader ../../../tools/custom-loader.mjs ./node.js "$@"
-node --experimental-modules --loader ../../../tools/custom-loader.mjs ./node.js "$@"
