--- shells/pipes-shell/web/deploy/source/index.html
@@ -1,39 +1,0 @@
-<!--
-@license
-Copyright (c) 2018 Google Inc. All rights reserved.
-This code may only be used under the BSD style license found at
-http://polymer.github.io/LICENSE.txt
-Code distributed by Google as part of this project is also
-subject to an additional IP rights grant found at
-http://polymer.github.io/PATENTS.txt
--->
-<!doctype html>
-<html lang="en">
-<head>
-
-  <title>Arcs</title>
-
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-
-  <style>
-    body {
-      margin: 0;
-      min-width: 320px;
-      overflow-x: hidden;
-      overflow-y: scroll;
-      background-color: white;
-    }
-    body, button, input {
-      font-family: 'Google Sans', sans-serif;
-      -webkit-font-smoothing: antialiased;
-    }
-  </style>
-
-  <script type="module" src="./shell.js"></script>
-
-</head>
-<body>
-
-</body>
-</html>
