--- shells/pipes-shell/web/deploy/stats.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-echo packing...
-npx webpack --profile --json > pack-stats.json
-echo studying...
-whybundled pack-stats.json
-
