--- shells/pipes-shell/web/index.html
@@ -1,29 +1,0 @@
-<!--
-@license
-Copyright (c) 2019 Google Inc. All rights reserved.
-This code may only be used under the BSD style license found at
-http://polymer.github.io/LICENSE.txt
-Code distributed by Google as part of this project is also
-subject to an additional IP rights grant found at
-http://polymer.github.io/PATENTS.txt
--->
-<!doctype html>
-<html lang="en">
-<head>
-
-  <title>Pipes::Arcs</title>
-
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-
-  <script src="../../configuration/csp.js"></script>
-  <!-- <link rel="stylesheet" href="../../configuration/index.css"> -->
-
-  <script type="module" src="./web.js"></script>
-
-</head>
-<body>
-
-</body>
-</html>
-
